pyproject.toml
[project]
name = "chromakit-backend"
version = "0.1.0"
description = "ChromaKit Backend: FastAPI service for image processing with Clean Architecture and Supabase integration"
authors = [
  { name = "ChromaKit Team" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "numpy>=1.26.0",
  "pillow>=10.3.0",
  "matplotlib>=3.8.0",
  "pydantic>=2.9.0",
  "supabase>=2.4.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "black>=24.4.2",
  "ruff>=0.5.0"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
testpaths = [
  "tests"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

